packages/evidentia-core/pyproject.toml
[project]
name = "evidentia-core"
version = "0.6.0"
description = "Core data models, catalog loading, and gap analysis engine for Evidentia GRC tool"
readme = "README.md"
authors = [{name = "Allen Byrd", email = "allen@allenfbyrd.com"}]
license = "Apache-2.0"
requires-python = ">=3.12"
keywords = ["grc", "compliance", "oscal", "nist", "soc2", "risk", "gap-analysis", "pydantic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "pyyaml>=6.0",
    "python-dateutil>=2.9",
    "thefuzz[speedup]>=0.22",
    "platformdirs>=4.3",
]

[project.urls]
Homepage = "https://github.com/allenfbyrd/evidentia"
Repository = "https://github.com/allenfbyrd/evidentia"
Issues = "https://github.com/allenfbyrd/evidentia/issues"
Changelog = "https://github.com/allenfbyrd/evidentia/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/evidentia_core"]
